New template + user unit test.
The difference is that feature_extactor will receive a row-wise list of dicts, not a dict of lists. So each element in the list if one instance (star), so you access individual features by X_dict[i]['asc_d']. In the simple version (where I'm only selecting a set of features from the static table, you construct the np array like this:

        cols = [
            'magnitude_b',
            'magnitude_r',
            'period',
            'asym_b',
            'asym_r',
            'log_p_not_variable',
            'flux_b',
            'flux_r',
            'quality',
            'div_period'
        ]
        return np.array([[instance[col] for col in cols] for instance in X_dict])

For Alex and Djalel: I had to go back to to_dict(orient='records') because it turns out that the splitting should be done before we call feature extractor. So our general template will be that feature_extractor will always receive a list of objects (and the objects will vary from RAMP to RAMP). Since X will not always be an np.array, we cannot call cross-validate directly on X, but we can use an index table (skf) writing down explicitly the indexing:

    for is in skf:
        X_cv = [X[i] for i in is]

The idea is that I can keep this scheme in generic, and specify what X is in specific (to RAMP).